+++ lib/stellar/order_books.ex
@@ -19,11 +19,12 @@ defmodule Stellar.OrderBooks do
 
   * `limit`: Maximum number of records to return.
   """
-  @spec get(Stellar.asset_type, Stellar.asset_type, Keyword.t) :: {Stellar.status, map}
+  @spec get(Stellar.asset_type(), Stellar.asset_type(), Keyword.t()) :: {Stellar.status(), map}
   def get(selling_asset_type, buying_asset_type, params \\ []) do
-    params = params
-    |> Keyword.put(:selling_asset_type, selling_asset_type)
-    |> Keyword.put(:buying_asset_type, buying_asset_type)
+    params =
+      params
+      |> Keyword.put(:selling_asset_type, selling_asset_type)
+      |> Keyword.put(:buying_asset_type, buying_asset_type)
 
     query = Base.process_query_params(params)
     Base.get("/order_book#{query}")