+++ test/stellar/ledgers_test.exs
@@ -3,17 +3,17 @@ defmodule Stellar.Ledgers.Test do
   alias Stellar.Ledgers
 
   test "get ledger details", %{bypass: bypass} do
-    Bypass.expect_once bypass, "GET", "/ledgers/123456", fn conn ->
+    Bypass.expect_once(bypass, "GET", "/ledgers/123456", fn conn ->
       Plug.Conn.resp(conn, 200, ~s<{"id": "123456"}>)
-    end
+    end)
 
     assert {:ok, %{"id" => "123456"}} = Ledgers.get("123456")
   end
 
   test "get all ledgers", %{bypass: bypass} do
-    Bypass.expect_once bypass, "GET", "/ledgers", fn conn ->
+    Bypass.expect_once(bypass, "GET", "/ledgers", fn conn ->
       Plug.Conn.resp(conn, 200, ~s<{"_embedded": { "records": [] }}>)
-    end
+    end)
 
     assert {:ok, %{"_embedded" => _}} = Ledgers.all()
   end