+++ test/stellar/order_books_test.exs
@@ -3,9 +3,9 @@ defmodule Stellar.OrderBooks.Test do
   alias Stellar.OrderBooks
 
   test "get order book details", %{bypass: bypass} do
-    Bypass.expect_once bypass, "GET", "/order_book", fn conn ->
+    Bypass.expect_once(bypass, "GET", "/order_book", fn conn ->
       Plug.Conn.resp(conn, 200, ~s<{"bids": []}>)
-    end
+    end)
 
     assert {:ok, %{"bids" => []}} = OrderBooks.get(:native, :native)
   end