+++ test/stellar/str_key_test.exs
@@ -9,8 +9,9 @@ defmodule Stellar.StrKey.Test do
   end
 
   test "encode_check! with invalid version_byte_name raises" do
-    data = <<85, 170, 163, 24, 149, 77, 255, 19, 245, 22, 224, 124, 2, 45, 42, 241, 162, 2,
-    23, 25, 226, 47, 133, 151, 8, 76, 130, 155, 252, 4, 95, 22>>
+    data =
+      <<85, 170, 163, 24, 149, 77, 255, 19, 245, 22, 224, 124, 2, 45, 42, 241, 162, 2, 23, 25,
+        226, 47, 133, 151, 8, 76, 130, 155, 252, 4, 95, 22>>
 
     assert_raise ArgumentError, fn ->
       StrKey.encode_check!(:something, data)
@@ -42,7 +43,7 @@ defmodule Stellar.StrKey.Test do
   end
 
   test "decode_check! to encode_check! and back" do
-    Enum.each(1..10, fn(_) ->
+    Enum.each(1..10, fn _ ->
       {public, secret} = KeyPair.random()
 
       data = StrKey.decode_check!(:ed25519SecretSeed, secret)