+++ test/stellar/trade_aggregations_test.exs
@@ -3,13 +3,14 @@ defmodule Stellar.TradeAggregations.Test do
   alias Stellar.TradeAggregations
 
   test "get all trade aggregrations", %{bypass: bypass} do
-    Bypass.expect_once bypass, "GET", "/trade_aggregations", fn conn ->
+    Bypass.expect_once(bypass, "GET", "/trade_aggregations", fn conn ->
       Plug.Conn.resp(conn, 200, ~s<{"_embedded": { "records": [] }}>)
-    end
+    end)
 
-    assert {:ok, %{"_embedded" => _}} = TradeAggregations.all(
-      :native,
-      :native
-    )
+    assert {:ok, %{"_embedded" => _}} =
+             TradeAggregations.all(
+               :native,
+               :native
+             )
   end
 end